Unpin python
Unfortunately, when the deadsnakes PPA updates python versions,
launchpad removes the previous version.
https://bitbucket.org/fk/deadsnakes-issues/issues/32/2710-missing-from-deadsnakes-27-ppa

We would love to pin, but getting all of the python packages installed using
gdebi or dpkg -i requires a bunch of fiddling.

If we find that keeping up with deadsnakes is untenable, we can stand up our
own S3 PPA to host old copies of the debs, or tell virtualenv to build python
from a tarball for each IDA.